Remove usage of double for hit and crit.
Use unsigned as it would show up in attack tables, i.e. 1/10000th.

+++ CombatRoll/AttackTables/MeleeWhiteHitTable.h
@@ -16,7 +16,7 @@ public:
                   const double block);
 
     int get_outcome(const unsigned roll,
-                    const double crit_chance,
+                    const unsigned crit_chance,
                     const bool include_dodge = true,
                     const bool include_parry = true,
                     const bool include_block = true,
