+++ CombatRoll/CombatRoll.h
@@ -28,9 +28,9 @@ public:
                                  const bool include_parry = true,
                                  const bool include_block = true,
                                  const bool include_miss = true);
-    int get_ranged_hit_result(const int wpn_skill, const double crit_chance);
-    int get_ranged_ability_result(const int wpn_skill, const double crit_chance);
-    int get_spell_ability_result(const MagicSchool, const double);
+    int get_ranged_hit_result(const int wpn_skill, const unsigned crit_chance);
+    int get_ranged_ability_result(const int wpn_skill, const unsigned crit_chance);
+    int get_spell_ability_result(const MagicSchool, const unsigned);
     int get_spell_resist_result(const MagicSchool);
 
     MeleeWhiteHitTable* get_melee_white_table(const int);
@@ -44,9 +44,9 @@ public:
     double get_yellow_miss_chance(const int wpn_skill);
     double get_glancing_blow_dmg_penalty(const int);
 
-    void update_melee_miss_chance(const double hit);
-    void update_ranged_miss_chance(const double hit);
-    void update_spell_miss_chance(const double spell_hit);
+    void update_melee_miss_chance(const unsigned hit);
+    void update_ranged_miss_chance(const unsigned hit);
+    void update_spell_miss_chance(const unsigned spell_hit);
 
     void drop_tables();
